C in CRUD: Create a record in the db and update content on the website asynchronously based on a positive response of the server.
Operations performed on the app (thanks Flask-SQLAlchemy):
flask db init
flask db migrate -m "Initial migration."
flask db upgrade

At this point assume that we add some records to the database.
The next migration that will be performed needs to take it into account
(hand-tuning the migration e.g. if there are constraints on adding the
column to the db).